refactor(v1): move v1 API to a separate package
